All needed files moved to VASTTesting.  Broker type created for TCP port connection.
