--- VAST/cpp/Obstacle.cpp
@@ -1,24 +1,0 @@
-#include "..\h\Obstacle.h"
-
-
-Obstacle::Obstacle(vector3 _position, vector3 _orientation, vector3 _dimensions)
-{
-
-}
-
-Obstacle::~Obstacle()
-{
-
-}
-
-void Obstacle::updateObstacle(vector3 _position, vector3 _orientation, vector3 _dimensions)
-{
-
-}
-
-void Obstacle::update()
-{
-
-}
-
-//Exten
