--- VAST/cpp/main.cpp
@@ -1,28 +1,0 @@
-//Virtual Autonomous System Testbed
-//ODU Capstone 2018-2019
-
-#include "..\h\VComponent.h"
-#include <iostream>
-#include <map>
-#include<variant>
-#include "..\h\VType.h"
-#include "..\h\AV.h"
-#include <Windows.h>
-
-
-int main(int argc, char **argv1)
-{
-	
-	return 0;
-}
-
-/*
-void sumoConnection() {
-	Client client;
-	client.connect("localhost", 1337);
-	std::cout << "time in ms: " << client.simulation.getCurrentTime() << "\n";
-	std::cout << "run 5 steps ...\n";
-	client.simulationStep(5 * 1000);
-	std::cout << "time in ms: " << client.simulation.getCurrentTime() << "\n";
-	client.close();
-} */
