+++ VASTTesting/Obstacle.cpp
@@ -1,0 +1,24 @@
+#include "Obstacle.h"
+
+
+Obstacle::Obstacle(Vector3 _position, Vector3 _orientation, Vector3 _dimensions)
+{
+
+}
+
+Obstacle::~Obstacle()
+{
+
+}
+
+void Obstacle::updateObstacle(Vector3 _position, Vector3 _orientation, Vector3 _dimensions)
+{
+
+}
+
+void Obstacle::update()
+{
+
+}
+
+//Exten
